better-dragscroll: Add option to activate dragging when host initial state is toggled
I tend to keep Num Lock always enabled while my keyboard is toggling
the value when I'm switching the scrolling key for my nano.

So instead of enabling the dragging phase depending on a fixed state of
a led, only care whether the led value had changed.

This also helps the case on hot-plug when a lock value is already set.